Update: unique, uniqueBy 함수 분리
