+++ src/4week/utils/unique.test.js
@@ -2,30 +2,12 @@
 const {I} = require('../internal_utils');
 
 // 어떤 것을 해볼까요?
-const unique = (arr, callback, isSorted) => {
+const unique = (arr) => {
   if (!I.isIterable(arr)) {
     return [];
   }
-  const map = new Map();
-  return arr.reduce((acc, item) => {
-    let key;
-    // Object인 경우
-    if (!Array.isArray(item)) {
-      key = JSON.stringify(item);
-    }
-    // callback이 있는 경우
-    if (typeof callback === 'function') {
-      key = callback(item);
-    }
-
-    if (!map.has(key)) {
-      acc.push(item);
-    }
-    map.set(key, item);
-
-    return acc;
-  }, []);
-};
+  return [...new Set(arr)];
+}
 
 describe('unique 테스트', () => {
   describe('non-lazy', () => {
@@ -56,21 +38,7 @@ describe('unique 테스트', () => {
       expect(thirdUniqueArray).toEqual([1, 2, 3]);
     });
 
-    it('case: 3, Advanced', () => {
-      const objects = [
-        { x: 1, y: 2 },
-        { x: 2, y: 1 },
-        { x: 1, y: 2 },
-      ];
-      const uniqueObjects = unique(objects);
-
-      expect(uniqueObjects).toEqual([
-        { x: 1, y: 2 },
-        { x: 2, y: 1 },
-      ]);
-    });
-
-    it('case: 4, arr is not iterable', () => {
+    it('case: 3, arr is not iterable', () => {
       const errorObjects = unique({ a: 'a', b: 'b' });
       
       expect(errorObjects).toEqual([]);
